Refactoring and bug fixing
1) Added Logger and Exception classes.
2) Added ShutdownMode for Application. Theme related classes changed.
3) SDL event handling has been changed.
They are now sorted by removal frequency. Events related to hiding and closing are now handled correctly.
4) Any and Delegate comparision update. Now delegate compares callable objects and not only their type.
Added MethodEventHandler to simplify registration method handlers.
5) Fixed bug in operator= which accept nullptr for Ref.
6) Removed 'Update" concept.
7) Attach/Dettach log logic tree now workds correcly. Window now gets attached/gettached on Show/Close.

+++ Sgl/Input/MouseAndKeyEventArgs.h
@@ -37,8 +37,8 @@ namespace Sgl
 	{
 		float X;
 		float Y;
-		float ScrolledByX;
-		float ScrolledByY;
+		int ScrolledByX;
+		int ScrolledByY;
 		MouseWheelDirection Direction;
 	};
 }
